feat : AdminProject Manager 기능 atomic 패턴으로 분리
