+++ src/components/atoms/EmptyStateMessage.vue
@@ -7,7 +7,9 @@ const props = defineProps<Props>()
 </script>
 
 <template>
-  <div class="rounded-lg border border-[#2a2a2a] bg-[#0f0f0f] px-3 py-4 text-center text-sm text-zinc-500">
+  <div
+    class="rounded-lg border border-[#2a2a2a] bg-[#0f0f0f] px-3 py-4 text-center text-sm text-zinc-500"
+  >
     {{ props.message }}
   </div>
 </template>
